+++ apps/app/page.tsx
@@ -1,6 +1,7 @@
 'use client'
 import { Button } from '@/components/Button';
 import { DatePicker } from '@/components/DatePicker';
+import { InnerLabelField } from '@/components/InnerLabelField';
 import { SlideDatePicker } from '@/components/SlideDatePicker';
 import { Controller, useForm } from 'react-hook-form';
 
@@ -16,6 +17,7 @@ export default function Page() {
         control={control}
         render={({ field }) => (
           <SlideDatePicker {...field} >
+            <InnerLabelField label='Birth Date' />
           </SlideDatePicker>
         )}
       />
@@ -27,6 +29,7 @@ export default function Page() {
           <DatePicker timePicker {...field} />
         )}
       />
+
 
       <Button theme='light' type='submit'>Submit</Button>
     </form>