+++ CHANGELOG.md
@@ -119,6 +119,7 @@ and fix tailwindcss issues.
 ### Added
 
 - `Calender` component added to the library.
+- `SlideDatePicker` component added to the library.
 
 ### Refactor